requirements.txt
astroid==1.6.3
autopep8==1.3.5
colorama==0.3.9
Django==2.0.4
djangorestframework==3.8.2
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.4.0
pylint==1.8.4
pylint-django==0.11
pylint-plugin-utils==0.2.6
pytz==2018.4
six==1.11.0
wrapt==1.10.11
